Added drawplayer class that painter uses <organisation
