+++ leaderboard.txt
@@ -1,2 +1,2 @@
-:0
+:860
